+++ src/lazy-observable.ts
@@ -1,5 +1,5 @@
 import {IDENTITY} from "./utils";
-import {observable, action} from "mobx";
+import {observable, extras, action} from "mobx";
 
 /**
  * `lazyObservable` creates an observable around a `fetch` method that will not be invoked
@@ -28,10 +28,10 @@ import {observable, action} from "mobx";
  *
  * @param {(sink: (newValue: T) => void) => void} fetch method that will be called the first time the value of this observable is accessed. The provided sink can be used to produce a new value, synchronously or asynchronously
  * @param {T} [initialValue=undefined] optional initialValue that will be returned from `current` as long as the `sink` has not been called at least once
- * @param {any} [modifier=IDENTITY] optional mobx modifier that determines the the comparison and recursion strategy of the observable, for example `asFlat` or `asStructure`
  * @returns {{
  *     current(): T,
- *     refresh(): T
+ *     refresh(): T,
+ *     reset(): T
  * }}
  */
 
@@ -41,24 +41,38 @@ export function lazyObservable<T>(
     modifier = IDENTITY
 ): {
     current(): T,
-    refresh(): T
+    refresh(): T,
+    reset(): T
 } {
     let started = false;
     const value = observable(modifier(initialValue));
     let currentFnc = () => {
         if (!started) {
             started = true;
-            fetch(action("lazyObservable-fetch", (newValue: T) => {
-                value.set(newValue);
-            }));
+            fetch((newValue: T) => {
+                extras.allowStateChanges(true, () => {
+                    value.set(newValue);
+                });
+            });
         }
         return value.get();
     };
+    let resetFnc = action("lazyObservable-reset", () => {
+      value.set(initialValue);
+      return value.get();
+    });
     return {
         current: currentFnc,
         refresh: () => {
-            started = false;
-            return currentFnc();
+            if (started) {
+                started = false;
+                return currentFnc();
+            } else {
+                return value.get();
+            }
+        },
+        reset: () => {
+            return resetFnc();
         }
     };
 }